deps/v8/test/webkit/dfg-inline-unused-this-method-check-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

This tests that inlining a function that does not use this does not result in this being lost entirely, if we succeed in doing method check optimizations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS baz(object, 0, 0) is 0
PASS baz(object, 1, 2) is 3
PASS baz(object, 2, 4) is 6
PASS baz(object, 3, 6) is 9
PASS baz(object, 4, 8) is 12
PASS baz(object, 5, 10) is 15
PASS baz(object, 6, 12) is 18
PASS baz(object, 7, 14) is 21
PASS baz(object, 8, 16) is 24
PASS baz(object, 9, 18) is 27
PASS baz(object, 10, 20) is 30
PASS baz(object, 11, 22) is 33
PASS baz(object, 12, 24) is 36
PASS baz(object, 13, 26) is 39
PASS baz(object, 14, 28) is 42
PASS baz(object, 15, 30) is 45
PASS baz(object, 16, 32) is 48
PASS baz(object, 17, 34) is 51
PASS baz(object, 18, 36) is 54
PASS baz(object, 19, 38) is 57
PASS baz(object, 20, 40) is 60
PASS baz(object, 21, 42) is 63
PASS baz(object, 22, 44) is 66
PASS baz(object, 23, 46) is 69
PASS baz(object, 24, 48) is 72
PASS baz(object, 25, 50) is 75
PASS baz(object, 26, 52) is 78
PASS baz(object, 27, 54) is 81
PASS baz(object, 28, 56) is 84
PASS baz(object, 29, 58) is 87
PASS baz(object, 30, 60) is 90
PASS baz(object, 31, 62) is 93
PASS baz(object, 32, 64) is 96
PASS baz(object, 33, 66) is 99
PASS baz(object, 34, 68) is 102
PASS baz(object, 35, 70) is 105
PASS baz(object, 36, 72) is 108
PASS baz(object, 37, 74) is 111
PASS baz(object, 38, 76) is 114
PASS baz(object, 39, 78) is 117
PASS baz(object, 40, 80) is 120
PASS baz(object, 41, 82) is 123
PASS baz(object, 42, 84) is 126
PASS baz(object, 43, 86) is 129
PASS baz(object, 44, 88) is 132
PASS baz(object, 45, 90) is 135
PASS baz(object, 46, 92) is 138
PASS baz(object, 47, 94) is 141
PASS baz(object, 48, 96) is 144
PASS baz(object, 49, 98) is 147
PASS baz(object, 50, 100) is 150
PASS baz(object, 51, 102) is 153
PASS baz(object, 52, 104) is 156
PASS baz(object, 53, 106) is 159
PASS baz(object, 54, 108) is 162
PASS baz(object, 55, 110) is 165
PASS baz(object, 56, 112) is 168
PASS baz(object, 57, 114) is 171
PASS baz(object, 58, 116) is 174
PASS baz(object, 59, 118) is 177
PASS baz(object, 60, 120) is 180
PASS baz(object, 61, 122) is 183
PASS baz(object, 62, 124) is 186
PASS baz(object, 63, 126) is 189
PASS baz(object, 64, 128) is 192
PASS baz(object, 65, 130) is 195
PASS baz(object, 66, 132) is 198
PASS baz(object, 67, 134) is 201
PASS baz(object, 68, 136) is 204
PASS baz(object, 69, 138) is 207
PASS baz(object, 70, 140) is 210
PASS baz(object, 71, 142) is 213
PASS baz(object, 72, 144) is 216
PASS baz(object, 73, 146) is 219
PASS baz(object, 74, 148) is 222
PASS baz(object, 75, 150) is 225
PASS baz(object, 76, 152) is 228
PASS baz(object, 77, 154) is 231
PASS baz(object, 78, 156) is 234
PASS baz(object, 79, 158) is 237
PASS baz(object, 80, 160) is 240
PASS baz(object, 81, 162) is 243
PASS baz(object, 82, 164) is 246
PASS baz(object, 83, 166) is 249
PASS baz(object, 84, 168) is 252
PASS baz(object, 85, 170) is 255
PASS baz(object, 86, 172) is 258
PASS baz(object, 87, 174) is 261
PASS baz(object, 88, 176) is 264
PASS baz(object, 89, 178) is 267
PASS baz(object, 90, 180) is 270
PASS baz(object, 91, 182) is 273
PASS baz(object, 92, 184) is 276
PASS baz(object, 93, 186) is 279
PASS baz(object, 94, 188) is 282
PASS baz(object, 95, 190) is 285
PASS baz(object, 96, 192) is 288
PASS baz(object, 97, 194) is 291
PASS baz(object, 98, 196) is 294
PASS baz(object, 99, 198) is 297
PASS baz(object, 100, 200) is 300
PASS baz(object, 101, 202) is 303
PASS baz(object, 102, 204) is 306
PASS baz(object, 103, 206) is 309
PASS baz(object, 104, 208) is 312
PASS baz(object, 105, 210) is 315
PASS baz(object, 106, 212) is 318
PASS baz(object, 107, 214) is 321
PASS baz(object, 108, 216) is 324
PASS baz(object, 109, 218) is 327
PASS baz(object, 110, 220) is 330
PASS baz(object, 111, 222) is 333
PASS baz(object, 112, 224) is 336
PASS baz(object, 113, 226) is 339
PASS baz(object, 114, 228) is 342
PASS baz(object, 115, 230) is 345
PASS baz(object, 116, 232) is 348
PASS baz(object, 117, 234) is 351
PASS baz(object, 118, 236) is 354
PASS baz(object, 119, 238) is 357
PASS baz(object, 120, 240) is 360
PASS baz(object, 121, 242) is 363
PASS baz(object, 122, 244) is 366
PASS baz(object, 123, 246) is 369
PASS baz(object, 124, 248) is 372
PASS baz(object, 125, 250) is 375
PASS baz(object, 126, 252) is 378
PASS baz(object, 127, 254) is 381
PASS baz(object, 128, 256) is 384
PASS baz(object, 129, 258) is 387
PASS baz(object, 130, 260) is 390
PASS baz(object, 131, 262) is 393
PASS baz(object, 132, 264) is 396
PASS baz(object, 133, 266) is 399
PASS baz(object, 134, 268) is 402
PASS baz(object, 135, 270) is 405
PASS baz(object, 136, 272) is 408
PASS baz(object, 137, 274) is 411
PASS baz(object, 138, 276) is 414
PASS baz(object, 139, 278) is 417
PASS baz(object, 140, 280) is 420
PASS baz(object, 141, 282) is 423
PASS baz(object, 142, 284) is 426
PASS baz(object, 143, 286) is 429
PASS baz(object, 144, 288) is 432
PASS baz(object, 145, 290) is 435
PASS baz(object, 146, 292) is 438
PASS baz(object, 147, 294) is 441
PASS baz(object, 148, 296) is 444
PASS baz(object, 149, 298) is 447
PASS baz(object, 150, 300) is 450
PASS baz(object, 151, 302) is 453
PASS baz(object, 152, 304) is 456
PASS baz(object, 153, 306) is 459
PASS baz(object, 154, 308) is 462
PASS baz(object, 155, 310) is 465
PASS baz(object, 156, 312) is 468
PASS baz(object, 157, 314) is 471
PASS baz(object, 158, 316) is 474
PASS baz(object, 159, 318) is 477
PASS baz(object, 160, 320) is 480
PASS baz(object, 161, 322) is 483
PASS baz(object, 162, 324) is 486
PASS baz(object, 163, 326) is 489
PASS baz(object, 164, 328) is 492
PASS baz(object, 165, 330) is 495
PASS baz(object, 166, 332) is 498
PASS baz(object, 167, 334) is 501
PASS baz(object, 168, 336) is 504
PASS baz(object, 169, 338) is 507
PASS baz(object, 170, 340) is 510
PASS baz(object, 171, 342) is 513
PASS baz(object, 172, 344) is 516
PASS baz(object, 173, 346) is 519
PASS baz(object, 174, 348) is 522
PASS baz(object, 175, 350) is 525
PASS baz(object, 176, 352) is 528
PASS baz(object, 177, 354) is 531
PASS baz(object, 178, 356) is 534
PASS baz(object, 179, 358) is 537
PASS baz(object, 180, 360) is 540
PASS baz(object, 181, 362) is 543
PASS baz(object, 182, 364) is 546
PASS baz(object, 183, 366) is 549
PASS baz(object, 184, 368) is 552
PASS baz(object, 185, 370) is 555
PASS baz(object, 186, 372) is 558
PASS baz(object, 187, 374) is 561
PASS baz(object, 188, 376) is 564
PASS baz(object, 189, 378) is 567
PASS baz(object, 190, 380) is 570
PASS baz(object, 191, 382) is 573
PASS baz(object, 192, 384) is 576
PASS baz(object, 193, 386) is 579
PASS baz(object, 194, 388) is 582
PASS baz(object, 195, 390) is 585
PASS baz(object, 196, 392) is 588
PASS baz(object, 197, 394) is 591
PASS baz(object, 198, 396) is 594
PASS baz(object, 199, 398) is 597
PASS baz(object, 200, 400) is 600
PASS baz(object, 201, 402) is 603
PASS baz(object, 202, 404) is 606
PASS baz(object, 203, 406) is 609
PASS baz(object, 204, 408) is 612
PASS baz(object, 205, 410) is 615
PASS baz(object, 206, 412) is 618
PASS baz(object, 207, 414) is 621
PASS baz(object, 208, 416) is 624
PASS baz(object, 209, 418) is 627
PASS baz(object, 210, 420) is 630
PASS baz(object, 211, 422) is 633
PASS baz(object, 212, 424) is 636
PASS baz(object, 213, 426) is 639
PASS baz(object, 214, 428) is 642
PASS baz(object, 215, 430) is 645
PASS baz(object, 216, 432) is 648
PASS baz(object, 217, 434) is 651
PASS baz(object, 218, 436) is 654
PASS baz(object, 219, 438) is 657
PASS baz(object, 220, 440) is 660
PASS baz(object, 221, 442) is 663
PASS baz(object, 222, 444) is 666
PASS baz(object, 223, 446) is 669
PASS baz(object, 224, 448) is 672
PASS baz(object, 225, 450) is 675
PASS baz(object, 226, 452) is 678
PASS baz(object, 227, 454) is 681
PASS baz(object, 228, 456) is 684
PASS baz(object, 229, 458) is 687
PASS baz(object, 230, 460) is 690
PASS baz(object, 231, 462) is 693
PASS baz(object, 232, 464) is 696
PASS baz(object, 233, 466) is 699
PASS baz(object, 234, 468) is 702
PASS baz(object, 235, 470) is 705
PASS baz(object, 236, 472) is 708
PASS baz(object, 237, 474) is 711
PASS baz(object, 238, 476) is 714
PASS baz(object, 239, 478) is 717
PASS baz(object, 240, 480) is 720
PASS baz(object, 241, 482) is 723
PASS baz(object, 242, 484) is 726
PASS baz(object, 243, 486) is 729
PASS baz(object, 244, 488) is 732
PASS baz(object, 245, 490) is 735
PASS baz(object, 246, 492) is 738
PASS baz(object, 247, 494) is 741
PASS baz(object, 248, 496) is 744
PASS baz(object, 249, 498) is 747
PASS baz(object, 250, 500) is 750
PASS baz(object, 251, 502) is 753
PASS baz(object, 252, 504) is 756
PASS baz(object, 253, 506) is 759
PASS baz(object, 254, 508) is 762
PASS baz(object, 255, 510) is 765
PASS baz(object, 256, 512) is 768
PASS baz(object, 257, 514) is 771
PASS baz(object, 258, 516) is 774
PASS baz(object, 259, 518) is 777
PASS baz(object, 260, 520) is 780
PASS baz(object, 261, 522) is 783
PASS baz(object, 262, 524) is 786
PASS baz(object, 263, 526) is 789
PASS baz(object, 264, 528) is 792
PASS baz(object, 265, 530) is 795
PASS baz(object, 266, 532) is 798
PASS baz(object, 267, 534) is 801
PASS baz(object, 268, 536) is 804
PASS baz(object, 269, 538) is 807
PASS baz(object, 270, 540) is 810
PASS baz(object, 271, 542) is 813
PASS baz(object, 272, 544) is 816
PASS baz(object, 273, 546) is 819
PASS baz(object, 274, 548) is 822
PASS baz(object, 275, 550) is 825
PASS baz(object, 276, 552) is 828
PASS baz(object, 277, 554) is 831
PASS baz(object, 278, 556) is 834
PASS baz(object, 279, 558) is 837
PASS baz(object, 280, 560) is 840
PASS baz(object, 281, 562) is 843
PASS baz(object, 282, 564) is 846
PASS baz(object, 283, 566) is 849
PASS baz(object, 284, 568) is 852
PASS baz(object, 285, 570) is 855
PASS baz(object, 286, 572) is 858
PASS baz(object, 287, 574) is 861
PASS baz(object, 288, 576) is 864
PASS baz(object, 289, 578) is 867
PASS baz(object, 290, 580) is 870
PASS baz(object, 291, 582) is 873
PASS baz(object, 292, 584) is 876
PASS baz(object, 293, 586) is 879
PASS baz(object, 294, 588) is 882
PASS baz(object, 295, 590) is 885
PASS baz(object, 296, 592) is 888
PASS baz(object, 297, 594) is 891
PASS baz(object, 298, 596) is 894
PASS baz(object, 299, 598) is 897
PASS baz(object, 300, 600) is 900
PASS baz(object, 301, 602) is 903
PASS baz(object, 302, 604) is 906
PASS baz(object, 303, 606) is 909
PASS baz(object, 304, 608) is 912
PASS baz(object, 305, 610) is 915
PASS baz(object, 306, 612) is 918
PASS baz(object, 307, 614) is 921
PASS baz(object, 308, 616) is 924
PASS baz(object, 309, 618) is 927
PASS baz(object, 310, 620) is 930
PASS baz(object, 311, 622) is 933
PASS baz(object, 312, 624) is 936
PASS baz(object, 313, 626) is 939
PASS baz(object, 314, 628) is 942
PASS baz(object, 315, 630) is 945
PASS baz(object, 316, 632) is 948
PASS baz(object, 317, 634) is 951
PASS baz(object, 318, 636) is 954
PASS baz(object, 319, 638) is 957
PASS baz(object, 320, 640) is 960
PASS baz(object, 321, 642) is 963
PASS baz(object, 322, 644) is 966
PASS baz(object, 323, 646) is 969
PASS baz(object, 324, 648) is 972
PASS baz(object, 325, 650) is 975
PASS baz(object, 326, 652) is 978
PASS baz(object, 327, 654) is 981
PASS baz(object, 328, 656) is 984
PASS baz(object, 329, 658) is 987
PASS baz(object, 330, 660) is 990
PASS baz(object, 331, 662) is 993
PASS baz(object, 332, 664) is 996
PASS baz(object, 333, 666) is 999
PASS baz(object, 334, 668) is 1002
PASS baz(object, 335, 670) is 1005
PASS baz(object, 336, 672) is 1008
PASS baz(object, 337, 674) is 1011
PASS baz(object, 338, 676) is 1014
PASS baz(object, 339, 678) is 1017
PASS baz(object, 340, 680) is 1020
PASS baz(object, 341, 682) is 1023
PASS baz(object, 342, 684) is 1026
PASS baz(object, 343, 686) is 1029
PASS baz(object, 344, 688) is 1032
PASS baz(object, 345, 690) is 1035
PASS baz(object, 346, 692) is 1038
PASS baz(object, 347, 694) is 1041
PASS baz(object, 348, 696) is 1044
PASS baz(object, 349, 698) is 1047
PASS baz(object, 350, 700) is 1050
PASS baz(object, 351, 702) is 1053
PASS baz(object, 352, 704) is 1056
PASS baz(object, 353, 706) is 1059
PASS baz(object, 354, 708) is 1062
PASS baz(object, 355, 710) is 1065
PASS baz(object, 356, 712) is 1068
PASS baz(object, 357, 714) is 1071
PASS baz(object, 358, 716) is 1074
PASS baz(object, 359, 718) is 1077
PASS baz(object, 360, 720) is 1080
PASS baz(object, 361, 722) is 1083
PASS baz(object, 362, 724) is 1086
PASS baz(object, 363, 726) is 1089
PASS baz(object, 364, 728) is 1092
PASS baz(object, 365, 730) is 1095
PASS baz(object, 366, 732) is 1098
PASS baz(object, 367, 734) is 1101
PASS baz(object, 368, 736) is 1104
PASS baz(object, 369, 738) is 1107
PASS baz(object, 370, 740) is 1110
PASS baz(object, 371, 742) is 1113
PASS baz(object, 372, 744) is 1116
PASS baz(object, 373, 746) is 1119
PASS baz(object, 374, 748) is 1122
PASS baz(object, 375, 750) is 1125
PASS baz(object, 376, 752) is 1128
PASS baz(object, 377, 754) is 1131
PASS baz(object, 378, 756) is 1134
PASS baz(object, 379, 758) is 1137
PASS baz(object, 380, 760) is 1140
PASS baz(object, 381, 762) is 1143
PASS baz(object, 382, 764) is 1146
PASS baz(object, 383, 766) is 1149
PASS baz(object, 384, 768) is 1152
PASS baz(object, 385, 770) is 1155
PASS baz(object, 386, 772) is 1158
PASS baz(object, 387, 774) is 1161
PASS baz(object, 388, 776) is 1164
PASS baz(object, 389, 778) is 1167
PASS baz(object, 390, 780) is 1170
PASS baz(object, 391, 782) is 1173
PASS baz(object, 392, 784) is 1176
PASS baz(object, 393, 786) is 1179
PASS baz(object, 394, 788) is 1182
PASS baz(object, 395, 790) is 1185
PASS baz(object, 396, 792) is 1188
PASS baz(object, 397, 794) is 1191
PASS baz(object, 398, 796) is 1194
PASS baz(object, 399, 798) is 1197
PASS baz(object, 400, 800) is 1200
PASS baz(object, 401, 802) is 1203
PASS baz(object, 402, 804) is 1206
PASS baz(object, 403, 806) is 1209
PASS baz(object, 404, 808) is 1212
PASS baz(object, 405, 810) is 1215
PASS baz(object, 406, 812) is 1218
PASS baz(object, 407, 814) is 1221
PASS baz(object, 408, 816) is 1224
PASS baz(object, 409, 818) is 1227
PASS baz(object, 410, 820) is 1230
PASS baz(object, 411, 822) is 1233
PASS baz(object, 412, 824) is 1236
PASS baz(object, 413, 826) is 1239
PASS baz(object, 414, 828) is 1242
PASS baz(object, 415, 830) is 1245
PASS baz(object, 416, 832) is 1248
PASS baz(object, 417, 834) is 1251
PASS baz(object, 418, 836) is 1254
PASS baz(object, 419, 838) is 1257
PASS baz(object, 420, 840) is 1260
PASS baz(object, 421, 842) is 1263
PASS baz(object, 422, 844) is 1266
PASS baz(object, 423, 846) is 1269
PASS baz(object, 424, 848) is 1272
PASS baz(object, 425, 850) is 1275
PASS baz(object, 426, 852) is 1278
PASS baz(object, 427, 854) is 1281
PASS baz(object, 428, 856) is 1284
PASS baz(object, 429, 858) is 1287
PASS baz(object, 430, 860) is 1290
PASS baz(object, 431, 862) is 1293
PASS baz(object, 432, 864) is 1296
PASS baz(object, 433, 866) is 1299
PASS baz(object, 434, 868) is 1302
PASS baz(object, 435, 870) is 1305
PASS baz(object, 436, 872) is 1308
PASS baz(object, 437, 874) is 1311
PASS baz(object, 438, 876) is 1314
PASS baz(object, 439, 878) is 1317
PASS baz(object, 440, 880) is 1320
PASS baz(object, 441, 882) is 1323
PASS baz(object, 442, 884) is 1326
PASS baz(object, 443, 886) is 1329
PASS baz(object, 444, 888) is 1332
PASS baz(object, 445, 890) is 1335
PASS baz(object, 446, 892) is 1338
PASS baz(object, 447, 894) is 1341
PASS baz(object, 448, 896) is 1344
PASS baz(object, 449, 898) is 1347
PASS baz(object, 450, 900) is 1350
PASS baz(object, 451, 902) is 1353
PASS baz(object, 452, 904) is 1356
PASS baz(object, 453, 906) is 1359
PASS baz(object, 454, 908) is 1362
PASS baz(object, 455, 910) is 1365
PASS baz(object, 456, 912) is 1368
PASS baz(object, 457, 914) is 1371
PASS baz(object, 458, 916) is 1374
PASS baz(object, 459, 918) is 1377
PASS baz(object, 460, 920) is 1380
PASS baz(object, 461, 922) is 1383
PASS baz(object, 462, 924) is 1386
PASS baz(object, 463, 926) is 1389
PASS baz(object, 464, 928) is 1392
PASS baz(object, 465, 930) is 1395
PASS baz(object, 466, 932) is 1398
PASS baz(object, 467, 934) is 1401
PASS baz(object, 468, 936) is 1404
PASS baz(object, 469, 938) is 1407
PASS baz(object, 470, 940) is 1410
PASS baz(object, 471, 942) is 1413
PASS baz(object, 472, 944) is 1416
PASS baz(object, 473, 946) is 1419
PASS baz(object, 474, 948) is 1422
PASS baz(object, 475, 950) is 1425
PASS baz(object, 476, 952) is 1428
PASS baz(object, 477, 954) is 1431
PASS baz(object, 478, 956) is 1434
PASS baz(object, 479, 958) is 1437
PASS baz(object, 480, 960) is 1440
PASS baz(object, 481, 962) is 1443
PASS baz(object, 482, 964) is 1446
PASS baz(object, 483, 966) is 1449
PASS baz(object, 484, 968) is 1452
PASS baz(object, 485, 970) is 1455
PASS baz(object, 486, 972) is 1458
PASS baz(object, 487, 974) is 1461
PASS baz(object, 488, 976) is 1464
PASS baz(object, 489, 978) is 1467
PASS baz(object, 490, 980) is 1470
PASS baz(object, 491, 982) is 1473
PASS baz(object, 492, 984) is 1476
PASS baz(object, 493, 986) is 1479
PASS baz(object, 494, 988) is 1482
PASS baz(object, 495, 990) is 1485
PASS baz(object, 496, 992) is 1488
PASS baz(object, 497, 994) is 1491
PASS baz(object, 498, 996) is 1494
PASS baz(object, 499, 998) is 1497
PASS baz(object, 500, 1000) is 1500
PASS baz(object, 501, 1002) is 1503
PASS baz(object, 502, 1004) is 1506
PASS baz(object, 503, 1006) is 1509
PASS baz(object, 504, 1008) is 1512
PASS baz(object, 505, 1010) is 1515
PASS baz(object, 506, 1012) is 1518
PASS baz(object, 507, 1014) is 1521
PASS baz(object, 508, 1016) is 1524
PASS baz(object, 509, 1018) is 1527
PASS baz(object, 510, 1020) is 1530
PASS baz(object, 511, 1022) is 1533
PASS baz(object, 512, 1024) is 1536
PASS baz(object, 513, 1026) is 1539
PASS baz(object, 514, 1028) is 1542
PASS baz(object, 515, 1030) is 1545
PASS baz(object, 516, 1032) is 1548
PASS baz(object, 517, 1034) is 1551
PASS baz(object, 518, 1036) is 1554
PASS baz(object, 519, 1038) is 1557
PASS baz(object, 520, 1040) is 1560
PASS baz(object, 521, 1042) is 1563
PASS baz(object, 522, 1044) is 1566
PASS baz(object, 523, 1046) is 1569
PASS baz(object, 524, 1048) is 1572
PASS baz(object, 525, 1050) is 1575
PASS baz(object, 526, 1052) is 1578
PASS baz(object, 527, 1054) is 1581
PASS baz(object, 528, 1056) is 1584
PASS baz(object, 529, 1058) is 1587
PASS baz(object, 530, 1060) is 1590
PASS baz(object, 531, 1062) is 1593
PASS baz(object, 532, 1064) is 1596
PASS baz(object, 533, 1066) is 1599
PASS baz(object, 534, 1068) is 1602
PASS baz(object, 535, 1070) is 1605
PASS baz(object, 536, 1072) is 1608
PASS baz(object, 537, 1074) is 1611
PASS baz(object, 538, 1076) is 1614
PASS baz(object, 539, 1078) is 1617
PASS baz(object, 540, 1080) is 1620
PASS baz(object, 541, 1082) is 1623
PASS baz(object, 542, 1084) is 1626
PASS baz(object, 543, 1086) is 1629
PASS baz(object, 544, 1088) is 1632
PASS baz(object, 545, 1090) is 1635
PASS baz(object, 546, 1092) is 1638
PASS baz(object, 547, 1094) is 1641
PASS baz(object, 548, 1096) is 1644
PASS baz(object, 549, 1098) is 1647
PASS baz(object, 550, 1100) is 1650
PASS baz(object, 551, 1102) is 1653
PASS baz(object, 552, 1104) is 1656
PASS baz(object, 553, 1106) is 1659
PASS baz(object, 554, 1108) is 1662
PASS baz(object, 555, 1110) is 1665
PASS baz(object, 556, 1112) is 1668
PASS baz(object, 557, 1114) is 1671
PASS baz(object, 558, 1116) is 1674
PASS baz(object, 559, 1118) is 1677
PASS baz(object, 560, 1120) is 1680
PASS baz(object, 561, 1122) is 1683
PASS baz(object, 562, 1124) is 1686
PASS baz(object, 563, 1126) is 1689
PASS baz(object, 564, 1128) is 1692
PASS baz(object, 565, 1130) is 1695
PASS baz(object, 566, 1132) is 1698
PASS baz(object, 567, 1134) is 1701
PASS baz(object, 568, 1136) is 1704
PASS baz(object, 569, 1138) is 1707
PASS baz(object, 570, 1140) is 1710
PASS baz(object, 571, 1142) is 1713
PASS baz(object, 572, 1144) is 1716
PASS baz(object, 573, 1146) is 1719
PASS baz(object, 574, 1148) is 1722
PASS baz(object, 575, 1150) is 1725
PASS baz(object, 576, 1152) is 1728
PASS baz(object, 577, 1154) is 1731
PASS baz(object, 578, 1156) is 1734
PASS baz(object, 579, 1158) is 1737
PASS baz(object, 580, 1160) is 1740
PASS baz(object, 581, 1162) is 1743
PASS baz(object, 582, 1164) is 1746
PASS baz(object, 583, 1166) is 1749
PASS baz(object, 584, 1168) is 1752
PASS baz(object, 585, 1170) is 1755
PASS baz(object, 586, 1172) is 1758
PASS baz(object, 587, 1174) is 1761
PASS baz(object, 588, 1176) is 1764
PASS baz(object, 589, 1178) is 1767
PASS baz(object, 590, 1180) is 1770
PASS baz(object, 591, 1182) is 1773
PASS baz(object, 592, 1184) is 1776
PASS baz(object, 593, 1186) is 1779
PASS baz(object, 594, 1188) is 1782
PASS baz(object, 595, 1190) is 1785
PASS baz(object, 596, 1192) is 1788
PASS baz(object, 597, 1194) is 1791
PASS baz(object, 598, 1196) is 1794
PASS baz(object, 599, 1198) is 1797
PASS baz(object, 600, 1200) is 1842
PASS baz(object, 601, 1202) is 1845
PASS baz(object, 602, 1204) is 1848
PASS baz(object, 603, 1206) is 1851
PASS baz(object, 604, 1208) is 1854
PASS baz(object, 605, 1210) is 1857
PASS baz(object, 606, 1212) is 1860
PASS baz(object, 607, 1214) is 1863
PASS baz(object, 608, 1216) is 1866
PASS baz(object, 609, 1218) is 1869
PASS baz(object, 610, 1220) is 1872
PASS baz(object, 611, 1222) is 1875
PASS baz(object, 612, 1224) is 1878
PASS baz(object, 613, 1226) is 1881
PASS baz(object, 614, 1228) is 1884
PASS baz(object, 615, 1230) is 1887
PASS baz(object, 616, 1232) is 1890
PASS baz(object, 617, 1234) is 1893
PASS baz(object, 618, 1236) is 1896
PASS baz(object, 619, 1238) is 1899
PASS baz(object, 620, 1240) is 1902
PASS baz(object, 621, 1242) is 1905
PASS baz(object, 622, 1244) is 1908
PASS baz(object, 623, 1246) is 1911
PASS baz(object, 624, 1248) is 1914
PASS baz(object, 625, 1250) is 1917
PASS baz(object, 626, 1252) is 1920
PASS baz(object, 627, 1254) is 1923
PASS baz(object, 628, 1256) is 1926
PASS baz(object, 629, 1258) is 1929
PASS baz(object, 630, 1260) is 1932
PASS baz(object, 631, 1262) is 1935
PASS baz(object, 632, 1264) is 1938
PASS baz(object, 633, 1266) is 1941
PASS baz(object, 634, 1268) is 1944
PASS baz(object, 635, 1270) is 1947
PASS baz(object, 636, 1272) is 1950
PASS baz(object, 637, 1274) is 1953
PASS baz(object, 638, 1276) is 1956
PASS baz(object, 639, 1278) is 1959
PASS baz(object, 640, 1280) is 1962
PASS baz(object, 641, 1282) is 1965
PASS baz(object, 642, 1284) is 1968
PASS baz(object, 643, 1286) is 1971
PASS baz(object, 644, 1288) is 1974
PASS baz(object, 645, 1290) is 1977
PASS baz(object, 646, 1292) is 1980
PASS baz(object, 647, 1294) is 1983
PASS baz(object, 648, 1296) is 1986
PASS baz(object, 649, 1298) is 1989
PASS baz(object, 650, 1300) is 1992
PASS baz(object, 651, 1302) is 1995
PASS baz(object, 652, 1304) is 1998
PASS baz(object, 653, 1306) is 2001
PASS baz(object, 654, 1308) is 2004
PASS baz(object, 655, 1310) is 2007
PASS baz(object, 656, 1312) is 2010
PASS baz(object, 657, 1314) is 2013
PASS baz(object, 658, 1316) is 2016
PASS baz(object, 659, 1318) is 2019
PASS baz(object, 660, 1320) is 2022
PASS baz(object, 661, 1322) is 2025
PASS baz(object, 662, 1324) is 2028
PASS baz(object, 663, 1326) is 2031
PASS baz(object, 664, 1328) is 2034
PASS baz(object, 665, 1330) is 2037
PASS baz(object, 666, 1332) is 2040
PASS baz(object, 667, 1334) is 2043
PASS baz(object, 668, 1336) is 2046
PASS baz(object, 669, 1338) is 2049
PASS baz(object, 670, 1340) is 2052
PASS baz(object, 671, 1342) is 2055
PASS baz(object, 672, 1344) is 2058
PASS baz(object, 673, 1346) is 2061
PASS baz(object, 674, 1348) is 2064
PASS baz(object, 675, 1350) is 2067
PASS baz(object, 676, 1352) is 2070
PASS baz(object, 677, 1354) is 2073
PASS baz(object, 678, 1356) is 2076
PASS baz(object, 679, 1358) is 2079
PASS baz(object, 680, 1360) is 2082
PASS baz(object, 681, 1362) is 2085
PASS baz(object, 682, 1364) is 2088
PASS baz(object, 683, 1366) is 2091
PASS baz(object, 684, 1368) is 2094
PASS baz(object, 685, 1370) is 2097
PASS baz(object, 686, 1372) is 2100
PASS baz(object, 687, 1374) is 2103
PASS baz(object, 688, 1376) is 2106
PASS baz(object, 689, 1378) is 2109
PASS baz(object, 690, 1380) is 2112
PASS baz(object, 691, 1382) is 2115
PASS baz(object, 692, 1384) is 2118
PASS baz(object, 693, 1386) is 2121
PASS baz(object, 694, 1388) is 2124
PASS baz(object, 695, 1390) is 2127
PASS baz(object, 696, 1392) is 2130
PASS baz(object, 697, 1394) is 2133
PASS baz(object, 698, 1396) is 2136
PASS baz(object, 699, 1398) is 2139
PASS baz(object, 700, 1400) is 2142
PASS baz(object, 701, 1402) is 2145
PASS baz(object, 702, 1404) is 2148
PASS baz(object, 703, 1406) is 2151
PASS baz(object, 704, 1408) is 2154
PASS baz(object, 705, 1410) is 2157
PASS baz(object, 706, 1412) is 2160
PASS baz(object, 707, 1414) is 2163
PASS baz(object, 708, 1416) is 2166
PASS baz(object, 709, 1418) is 2169
PASS baz(object, 710, 1420) is 2172
PASS baz(object, 711, 1422) is 2175
PASS baz(object, 712, 1424) is 2178
PASS baz(object, 713, 1426) is 2181
PASS baz(object, 714, 1428) is 2184
PASS baz(object, 715, 1430) is 2187
PASS baz(object, 716, 1432) is 2190
PASS baz(object, 717, 1434) is 2193
PASS baz(object, 718, 1436) is 2196
PASS baz(object, 719, 1438) is 2199
PASS baz(object, 720, 1440) is 2202
PASS baz(object, 721, 1442) is 2205
PASS baz(object, 722, 1444) is 2208
PASS baz(object, 723, 1446) is 2211
PASS baz(object, 724, 1448) is 2214
PASS baz(object, 725, 1450) is 2217
PASS baz(object, 726, 1452) is 2220
PASS baz(object, 727, 1454) is 2223
PASS baz(object, 728, 1456) is 2226
PASS baz(object, 729, 1458) is 2229
PASS baz(object, 730, 1460) is 2232
PASS baz(object, 731, 1462) is 2235
PASS baz(object, 732, 1464) is 2238
PASS baz(object, 733, 1466) is 2241
PASS baz(object, 734, 1468) is 2244
PASS baz(object, 735, 1470) is 2247
PASS baz(object, 736, 1472) is 2250
PASS baz(object, 737, 1474) is 2253
PASS baz(object, 738, 1476) is 2256
PASS baz(object, 739, 1478) is 2259
PASS baz(object, 740, 1480) is 2262
PASS baz(object, 741, 1482) is 2265
PASS baz(object, 742, 1484) is 2268
PASS baz(object, 743, 1486) is 2271
PASS baz(object, 744, 1488) is 2274
PASS baz(object, 745, 1490) is 2277
PASS baz(object, 746, 1492) is 2280
PASS baz(object, 747, 1494) is 2283
PASS baz(object, 748, 1496) is 2286
PASS baz(object, 749, 1498) is 2289
PASS baz(object, 750, 1500) is 2292
PASS baz(object, 751, 1502) is 2295
PASS baz(object, 752, 1504) is 2298
PASS baz(object, 753, 1506) is 2301
PASS baz(object, 754, 1508) is 2304
PASS baz(object, 755, 1510) is 2307
PASS baz(object, 756, 1512) is 2310
PASS baz(object, 757, 1514) is 2313
PASS baz(object, 758, 1516) is 2316
PASS baz(object, 759, 1518) is 2319
PASS baz(object, 760, 1520) is 2322
PASS baz(object, 761, 1522) is 2325
PASS baz(object, 762, 1524) is 2328
PASS baz(object, 763, 1526) is 2331
PASS baz(object, 764, 1528) is 2334
PASS baz(object, 765, 1530) is 2337
PASS baz(object, 766, 1532) is 2340
PASS baz(object, 767, 1534) is 2343
PASS baz(object, 768, 1536) is 2346
PASS baz(object, 769, 1538) is 2349
PASS baz(object, 770, 1540) is 2352
PASS baz(object, 771, 1542) is 2355
PASS baz(object, 772, 1544) is 2358
PASS baz(object, 773, 1546) is 2361
PASS baz(object, 774, 1548) is 2364
PASS baz(object, 775, 1550) is 2367
PASS baz(object, 776, 1552) is 2370
PASS baz(object, 777, 1554) is 2373
PASS baz(object, 778, 1556) is 2376
PASS baz(object, 779, 1558) is 2379
PASS baz(object, 780, 1560) is 2382
PASS baz(object, 781, 1562) is 2385
PASS baz(object, 782, 1564) is 2388
PASS baz(object, 783, 1566) is 2391
PASS baz(object, 784, 1568) is 2394
PASS baz(object, 785, 1570) is 2397
PASS baz(object, 786, 1572) is 2400
PASS baz(object, 787, 1574) is 2403
PASS baz(object, 788, 1576) is 2406
PASS baz(object, 789, 1578) is 2409
PASS baz(object, 790, 1580) is 2412
PASS baz(object, 791, 1582) is 2415
PASS baz(object, 792, 1584) is 2418
PASS baz(object, 793, 1586) is 2421
PASS baz(object, 794, 1588) is 2424
PASS baz(object, 795, 1590) is 2427
PASS baz(object, 796, 1592) is 2430
PASS baz(object, 797, 1594) is 2433
PASS baz(object, 798, 1596) is 2436
PASS baz(object, 799, 1598) is 2439
PASS baz(object, 800, 1600) is 2442
PASS baz(object, 801, 1602) is 2445
PASS baz(object, 802, 1604) is 2448
PASS baz(object, 803, 1606) is 2451
PASS baz(object, 804, 1608) is 2454
PASS baz(object, 805, 1610) is 2457
PASS baz(object, 806, 1612) is 2460
PASS baz(object, 807, 1614) is 2463
PASS baz(object, 808, 1616) is 2466
PASS baz(object, 809, 1618) is 2469
PASS baz(object, 810, 1620) is 2472
PASS baz(object, 811, 1622) is 2475
PASS baz(object, 812, 1624) is 2478
PASS baz(object, 813, 1626) is 2481
PASS baz(object, 814, 1628) is 2484
PASS baz(object, 815, 1630) is 2487
PASS baz(object, 816, 1632) is 2490
PASS baz(object, 817, 1634) is 2493
PASS baz(object, 818, 1636) is 2496
PASS baz(object, 819, 1638) is 2499
PASS baz(object, 820, 1640) is 2502
PASS baz(object, 821, 1642) is 2505
PASS baz(object, 822, 1644) is 2508
PASS baz(object, 823, 1646) is 2511
PASS baz(object, 824, 1648) is 2514
PASS baz(object, 825, 1650) is 2517
PASS baz(object, 826, 1652) is 2520
PASS baz(object, 827, 1654) is 2523
PASS baz(object, 828, 1656) is 2526
PASS baz(object, 829, 1658) is 2529
PASS baz(object, 830, 1660) is 2532
PASS baz(object, 831, 1662) is 2535
PASS baz(object, 832, 1664) is 2538
PASS baz(object, 833, 1666) is 2541
PASS baz(object, 834, 1668) is 2544
PASS baz(object, 835, 1670) is 2547
PASS baz(object, 836, 1672) is 2550
PASS baz(object, 837, 1674) is 2553
PASS baz(object, 838, 1676) is 2556
PASS baz(object, 839, 1678) is 2559
PASS baz(object, 840, 1680) is 2562
PASS baz(object, 841, 1682) is 2565
PASS baz(object, 842, 1684) is 2568
PASS baz(object, 843, 1686) is 2571
PASS baz(object, 844, 1688) is 2574
PASS baz(object, 845, 1690) is 2577
PASS baz(object, 846, 1692) is 2580
PASS baz(object, 847, 1694) is 2583
PASS baz(object, 848, 1696) is 2586
PASS baz(object, 849, 1698) is 2589
PASS baz(object, 850, 1700) is 2592
PASS baz(object, 851, 1702) is 2595
PASS baz(object, 852, 1704) is 2598
PASS baz(object, 853, 1706) is 2601
PASS baz(object, 854, 1708) is 2604
PASS baz(object, 855, 1710) is 2607
PASS baz(object, 856, 1712) is 2610
PASS baz(object, 857, 1714) is 2613
PASS baz(object, 858, 1716) is 2616
PASS baz(object, 859, 1718) is 2619
PASS baz(object, 860, 1720) is 2622
PASS baz(object, 861, 1722) is 2625
PASS baz(object, 862, 1724) is 2628
PASS baz(object, 863, 1726) is 2631
PASS baz(object, 864, 1728) is 2634
PASS baz(object, 865, 1730) is 2637
PASS baz(object, 866, 1732) is 2640
PASS baz(object, 867, 1734) is 2643
PASS baz(object, 868, 1736) is 2646
PASS baz(object, 869, 1738) is 2649
PASS baz(object, 870, 1740) is 2652
PASS baz(object, 871, 1742) is 2655
PASS baz(object, 872, 1744) is 2658
PASS baz(object, 873, 1746) is 2661
PASS baz(object, 874, 1748) is 2664
PASS baz(object, 875, 1750) is 2667
PASS baz(object, 876, 1752) is 2670
PASS baz(object, 877, 1754) is 2673
PASS baz(object, 878, 1756) is 2676
PASS baz(object, 879, 1758) is 2679
PASS baz(object, 880, 1760) is 2682
PASS baz(object, 881, 1762) is 2685
PASS baz(object, 882, 1764) is 2688
PASS baz(object, 883, 1766) is 2691
PASS baz(object, 884, 1768) is 2694
PASS baz(object, 885, 1770) is 2697
PASS baz(object, 886, 1772) is 2700
PASS baz(object, 887, 1774) is 2703
PASS baz(object, 888, 1776) is 2706
PASS baz(object, 889, 1778) is 2709
PASS baz(object, 890, 1780) is 2712
PASS baz(object, 891, 1782) is 2715
PASS baz(object, 892, 1784) is 2718
PASS baz(object, 893, 1786) is 2721
PASS baz(object, 894, 1788) is 2724
PASS baz(object, 895, 1790) is 2727
PASS baz(object, 896, 1792) is 2730
PASS baz(object, 897, 1794) is 2733
PASS baz(object, 898, 1796) is 2736
PASS baz(object, 899, 1798) is 2739
PASS baz(object, 900, 1800) is 2742
PASS baz(object, 901, 1802) is 2745
PASS baz(object, 902, 1804) is 2748
PASS baz(object, 903, 1806) is 2751
PASS baz(object, 904, 1808) is 2754
PASS baz(object, 905, 1810) is 2757
PASS baz(object, 906, 1812) is 2760
PASS baz(object, 907, 1814) is 2763
PASS baz(object, 908, 1816) is 2766
PASS baz(object, 909, 1818) is 2769
PASS baz(object, 910, 1820) is 2772
PASS baz(object, 911, 1822) is 2775
PASS baz(object, 912, 1824) is 2778
PASS baz(object, 913, 1826) is 2781
PASS baz(object, 914, 1828) is 2784
PASS baz(object, 915, 1830) is 2787
PASS baz(object, 916, 1832) is 2790
PASS baz(object, 917, 1834) is 2793
PASS baz(object, 918, 1836) is 2796
PASS baz(object, 919, 1838) is 2799
PASS baz(object, 920, 1840) is 2802
PASS baz(object, 921, 1842) is 2805
PASS baz(object, 922, 1844) is 2808
PASS baz(object, 923, 1846) is 2811
PASS baz(object, 924, 1848) is 2814
PASS baz(object, 925, 1850) is 2817
PASS baz(object, 926, 1852) is 2820
PASS baz(object, 927, 1854) is 2823
PASS baz(object, 928, 1856) is 2826
PASS baz(object, 929, 1858) is 2829
PASS baz(object, 930, 1860) is 2832
PASS baz(object, 931, 1862) is 2835
PASS baz(object, 932, 1864) is 2838
PASS baz(object, 933, 1866) is 2841
PASS baz(object, 934, 1868) is 2844
PASS baz(object, 935, 1870) is 2847
PASS baz(object, 936, 1872) is 2850
PASS baz(object, 937, 1874) is 2853
PASS baz(object, 938, 1876) is 2856
PASS baz(object, 939, 1878) is 2859
PASS baz(object, 940, 1880) is 2862
PASS baz(object, 941, 1882) is 2865
PASS baz(object, 942, 1884) is 2868
PASS baz(object, 943, 1886) is 2871
PASS baz(object, 944, 1888) is 2874
PASS baz(object, 945, 1890) is 2877
PASS baz(object, 946, 1892) is 2880
PASS baz(object, 947, 1894) is 2883
PASS baz(object, 948, 1896) is 2886
PASS baz(object, 949, 1898) is 2889
PASS baz(object, 950, 1900) is 2892
PASS baz(object, 951, 1902) is 2895
PASS baz(object, 952, 1904) is 2898
PASS baz(object, 953, 1906) is 2901
PASS baz(object, 954, 1908) is 2904
PASS baz(object, 955, 1910) is 2907
PASS baz(object, 956, 1912) is 2910
PASS baz(object, 957, 1914) is 2913
PASS baz(object, 958, 1916) is 2916
PASS baz(object, 959, 1918) is 2919
PASS baz(object, 960, 1920) is 2922
PASS baz(object, 961, 1922) is 2925
PASS baz(object, 962, 1924) is 2928
PASS baz(object, 963, 1926) is 2931
PASS baz(object, 964, 1928) is 2934
PASS baz(object, 965, 1930) is 2937
PASS baz(object, 966, 1932) is 2940
PASS baz(object, 967, 1934) is 2943
PASS baz(object, 968, 1936) is 2946
PASS baz(object, 969, 1938) is 2949
PASS baz(object, 970, 1940) is 2952
PASS baz(object, 971, 1942) is 2955
PASS baz(object, 972, 1944) is 2958
PASS baz(object, 973, 1946) is 2961
PASS baz(object, 974, 1948) is 2964
PASS baz(object, 975, 1950) is 2967
PASS baz(object, 976, 1952) is 2970
PASS baz(object, 977, 1954) is 2973
PASS baz(object, 978, 1956) is 2976
PASS baz(object, 979, 1958) is 2979
PASS baz(object, 980, 1960) is 2982
PASS baz(object, 981, 1962) is 2985
PASS baz(object, 982, 1964) is 2988
PASS baz(object, 983, 1966) is 2991
PASS baz(object, 984, 1968) is 2994
PASS baz(object, 985, 1970) is 2997
PASS baz(object, 986, 1972) is 3000
PASS baz(object, 987, 1974) is 3003
PASS baz(object, 988, 1976) is 3006
PASS baz(object, 989, 1978) is 3009
PASS baz(object, 990, 1980) is 3012
PASS baz(object, 991, 1982) is 3015
PASS baz(object, 992, 1984) is 3018
PASS baz(object, 993, 1986) is 3021
PASS baz(object, 994, 1988) is 3024
PASS baz(object, 995, 1990) is 3027
PASS baz(object, 996, 1992) is 3030
PASS baz(object, 997, 1994) is 3033
PASS baz(object, 998, 1996) is 3036
PASS baz(object, 999, 1998) is 3039
PASS successfullyParsed is true

TEST COMPLETE
